+++ Runtime/ViewModels/ButtonListViewModel.cs
@@ -1,8 +1,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using Kuuasema.DataBinding;
-namespace Kuuasema.Core {
+namespace Kuuasema.DataBinding {
     public class ButtonListViewModel : ListViewModel<string, DataModel<string>, ButtonViewModel> {
 
     }