+++ Runtime/ViewModels/TextListViewModel.cs
@@ -1,4 +1,3 @@
-using Kuuasema.DataBinding;
-namespace Kuuasema.Core {
+namespace Kuuasema.DataBinding {
     public class TextListViewModel : ListViewModel<string, DataModel<string>, TextViewModel> { }
 }